Time: 215 ms (100%), Space: 109.7 MB (100%) - LeetHub

+++ 3485-maximize-score-of-numbers-in-ranges/3485-maximize-score-of-numbers-in-ranges.cpp
@@ -23,8 +23,7 @@ public:
         int ans = -1;
 
         while(low <= high){
-            // int mid = low + (high-low)/2;
-            long long mid = low + (high-low)/2;
+            int mid = low + (high-low)/2;
             
             if(isOkay(mid, start, d)){
                 ans = mid;
